refactor(permissions): rename macos_perms/ to permissions/ (#597)
Item 4 of the architecture-cleanup roadmap. Pure rename — no behaviour
change. Pre-emptively unblocks Linux (#106) and Windows (#107)
permission impls by giving cross-platform permission code a sensible
home.

Changes:
- src/macos_perms/ -> src/permissions/, with the macOS FFI extracted
  into permissions/macos.rs so the module name now reflects the
  abstraction rather than the only implementation.
- src/ipc/commands/macos.rs -> src/ipc/commands/permissions.rs (file
  rename via git mv; IPC command names unchanged because they're a
  wire-protocol contract with the frontend).
- All `crate::macos_perms::*` references updated (3 files, 13 refs).
- generate_handler! registrations in lib.rs updated.
- Doc comments updated to reflect the rename and the cross-platform
  rationale.

The module name was the only obstacle — the actual cfg-gating story
inside is unchanged. When Linux/Windows permission impls land they
become peers of permissions/macos.rs (e.g., permissions/linux.rs,
permissions/windows.rs) and the public API in permissions/mod.rs
dispatches per platform.

Verified locally:
- cargo check --all-targets clean
- cargo clippy --all-targets -- -D warnings clean
- cargo clippy --lib --no-default-features -- -D warnings clean
  (the cross-platform path from CLAUDE.md)
- cargo test --lib: 412 pass, 0 fail (same count as before)
